requirements.txt
praw==7.7.1
elevenlabs==0.2.26
moviepy==1.0.3
openai-whisper==20231117
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
ffmpeg-python==0.2.0 